requirements.txt
cachetools==2.0.1
certifi==2017.11.5
chardet==3.0.4
enum34==1.1.6
futures==3.2.0
google-api-core==0.1.4
google-auth==1.3.0
google-cloud-vision==0.29.0
googleapis-common-protos==1.5.3
grpcio==1.8.4
idna==2.6
protobuf==3.5.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pytz==2017.3
requests==2.18.4
rsa==3.4.2
six==1.11.0
urllib3==1.22
